requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.1.0
python-dotenv==1.0.0
pydantic==1.10.11
requests==2.31.0
openai==1.11.0       
mistral-inference==0.1.0
